recipes-core: Renamed and updated owswaas daemon launcher
The rename was a quick fix for the fact that "killall owswaas"
tanked both the running program and our service script. (Whoops!)

More GPIO setup has been moved into this script to allow the owswaas
process to run without significant privileges/capabilities. This
recipe is now responsible for creating the "owswaas" user.